Update install.rst
Just nice to have the reminder in the documentation as to how to activate the pipeline. We have a usamriidPathDiscov directory nested within usamriidPathDiscov directory which can get confusing as users will attempt to source from the higher directory and wonder why they can't find /bin/activate - so I added this code block.